Modify `GPIOTEST_dog.go` to not close channel
Previously `cancelButtonWav` was created and closed in
`delayedButtonHandle`. Unfortunately because its parent `pushButton` is
a shared struct, it was possible for `cancelButtonWav` to get closed
multiple times, across multiple invocations of `delayedButtonHandle`.

Create `cancelButtonWav` as a buffered channel, and instead send a
message to it rather than rely on closing.

It's possible some messages could leak through to
`pushButton.buttonWavs.Play(catcat)`, may have to validate this.

Signed-off-by: Andrew Seigner <andrew@sig.gy>